+++ tests/phpunit/test-uninstall.php
@@ -8,12 +8,15 @@
  */
 
 /**
- * Tests uninstalling the module.
+ * Tests uninstalling the extension.
  *
  * @since 1.0.0
+ *
+ * @covers WordPoints_Dynamic_Points_Installable
+ * @covers WordPoints_Dynamic_Points_Uninstaller_Reactions
  */
 class WordPoints_Dynamic_Points_Uninstall_Test
-	extends WordPoints_PHPUnit_TestCase_Module_Uninstall {
+	extends WordPoints_PHPUnit_TestCase_Extension_Uninstall {
 
 	/**
 	 * Test installation and uninstallation.
@@ -24,10 +27,10 @@ class WordPoints_Dynamic_Points_Uninstall_Test
 
 		$this->uninstall();
 
-		// Check that everything with this module's prefix has been uninstalled.
+		// Check that everything with this extension's prefix has been uninstalled.
 		$this->assertUninstalledPrefix( 'dynamic_points' );
 
-		if ( ! getenv( 'WORDPOINTS_ONLY_UNINSTALL_MODULE' ) ) {
+		if ( ! getenv( 'WORDPOINTS_ONLY_UNINSTALL_EXTENSION' ) ) {
 			return;
 		}