+++ CHANGELOG.md
@@ -8,6 +8,17 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a CH
 
 Not documented yet.
 
+## [1.0.1] - 2017-10-06
+
+### Requires
+
+- WordPoints: 2.4+
+
+### Fixed
+
+- Deprecated notices from `Channel`, `Module Name`, and `Module URI` extension headers.
+- Deprecated notices during uninstall, by using the new installables API.
+
 ## [1.0.0] - 2017-03-14
 
 ### Added
@@ -22,4 +33,5 @@ Not documented yet.
   - On uninstall, any reactions still using the extension are disabled.
 
 [unreleased]: https://github.com/WordPoints/wordpoints/compare/master...HEAD
+[1.0.1]: https://github.com/WordPoints/wordpoints/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/WordPoints/wordpoints/compare/...1.0.0